+++ src/fair_and_square/src/library/levels/showcase/DotShowcase.cc
@@ -15,7 +15,7 @@
 
 using namespace octopus;
 
-std::list<Steppable *> dotShowcaseSteps(Library &lib_p, size_t number_p)
+std::list<Steppable *> dotShowcaseSteps(Library &lib_p, uint32_t number_p)
 {
     // square
 {
@@ -61,7 +61,7 @@ std::list<Steppable *> dotShowcaseSteps(Library &lib_p, size_t number_p)
 
 	unsigned long id_l = 0;
     spawners_l.push_back(new UnitSpawnStep(Handle(id_l++), square1_l));
-	for(size_t i = 0; i < number_p ; ++ i)
+	for(uint32_t i = 0; i < number_p ; ++ i)
 	{
 		spawners_l.push_back(new UnitSpawnStep(Handle(id_l++), triangle2_l));
 	}
