+++ src/fair_and_square/src/library/model/AnchorDivinityTrigger.hh
@@ -34,7 +34,7 @@ private:
 
 	/// @brief count of trigger during the game to generate option key
 	/// need to be mutable to be incremented when triggering
-	mutable size_t _count {0};
+	mutable uint32_t _count {0};
 };
 
 #endif
